🚚move:캐러셀 컴포넌트 파일 이동
